Changed SceneManager to also add in the stuff there, to avoid having to use more notations to get through the functions.
