+++ src/animation/AnimPanes.java
@@ -1,0 +1,10 @@
+package animation;
+
+/** 
+ * Static-type class, where we can set-up pre-defined animations. 
+ * @author Johnny Nguyen
+ *
+ */
+public final class AnimPanes {
+
+}
