+++ src/example/Initialize.java
@@ -1,4 +1,4 @@
-package screenmanager;
+package example;
 
 import javafx.application.Application;
 import javafx.geometry.Pos;
@@ -6,6 +6,11 @@ import javafx.scene.control.Button;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
 import javafx.stage.Stage;
+import screenmanager.SceneManager;
+
+/*
+ * This class is supposed to show an example screen for you
+ */
 
 public class Initialize extends Application {
 
@@ -15,21 +20,21 @@ public class Initialize extends Application {
 		SceneManager scene = SceneManager.getInstance(800, 600);
 		// random Button to show-case switching screens
 		Button btn1 = new Button("Switch Screen");
-		btn1.setOnAction(e -> scene.getScreenManager().switchScreen("Pane 2"));
+		btn1.setOnAction(e -> scene.switchScreen("Pane 2"));
 		Button btn2 = new Button("Switch Screen");
-		btn2.setOnAction(e -> scene.getScreenManager().switchScreen("Pane 1"));
+		btn2.setOnAction(e -> scene.switchScreen("Pane 1"));
 		// Random Pane
 		VBox pane = new VBox();
 		VBox pane2 = new VBox();
-		pane.setStyle("-fx-background-color: blue;");
+		pane.setStyle("-fx-background-color: lightblue;");
 		pane2.setStyle("-fx-background-color: yellow;");
 		pane.getChildren().addAll(new Text("Pane 1"), btn1);
 		pane2.getChildren().addAll(new Text("Pane 2"), btn2);
 		pane.setAlignment(Pos.CENTER);
 		pane2.setAlignment(Pos.CENTER);
 		// add screen
-		scene.getScreenManager().addScreen("Pane 1", pane);
-		scene.getScreenManager().addScreen("Pane 2", pane2);
+		scene.addScreen("Pane 1", pane);
+		scene.addScreen("Pane 2", pane2);
 		// create stage
 		primaryStage.setScene(scene);
 		primaryStage.setResizable(false);
